+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - Improved error message when no concrete binded class was found
 
+## 0.5.0
+### 2023-05-19
+
+- Accept `Closure|string` in `getParametersToResolve()`
+- Added `resolve(callable)` to `ContainerInterface`
+
 ## 0.4.0
 ### 2023-04-27